refactor: Move `ClientReceiveState` into own file
