+++ src/lib.rs
@@ -1,6 +1,7 @@
 #![forbid(unsafe_code)]
 
 pub mod client;
+mod client_receive;
 mod handle;
 mod receive;
 mod send_command;
